do for all dialects
