+++ cryptography/src/bls12381/benches/partial_signature_aggregation.rs
@@ -8,24 +8,21 @@ fn benchmark_partial_signature_aggregation(c: &mut Criterion) {
     let msg = b"hello";
     for &n in &[5, 10, 20, 50, 100, 250, 500] {
         let t = quorum(n).unwrap();
-        c.bench_function(
-            &format!("{} partial_aggregate: n={} t={}", module_path!(), n, t),
-            |b| {
-                b.iter_batched(
-                    || {
-                        let (_, shares) = dkg::ops::generate_shares(None, n, t);
-                        shares
-                            .iter()
-                            .map(|s| primitives::ops::partial_sign(s, Some(namespace), msg))
-                            .collect::<Vec<_>>()
-                    },
-                    |partials| {
-                        black_box(primitives::ops::partial_aggregate(t, partials).unwrap());
-                    },
-                    BatchSize::SmallInput,
-                );
-            },
-        );
+        c.bench_function(&format!("{}/n={} t={}", module_path!(), n, t), |b| {
+            b.iter_batched(
+                || {
+                    let (_, shares) = dkg::ops::generate_shares(None, n, t);
+                    shares
+                        .iter()
+                        .map(|s| primitives::ops::partial_sign(s, Some(namespace), msg))
+                        .collect::<Vec<_>>()
+                },
+                |partials| {
+                    black_box(primitives::ops::partial_aggregate(t, partials).unwrap());
+                },
+                BatchSize::SmallInput,
+            );
+        });
     }
 }
