+++ cryptography/src/bls12381/benches/signature_aggregation.rs
@@ -12,26 +12,23 @@ fn benchmark_signature_aggregation(c: &mut Criterion) {
             thread_rng().fill(&mut msg);
             msgs.push(msg);
         }
-        c.bench_function(
-            &format!("{} aggregate: msgs={}", module_path!(), msgs.len()),
-            |b| {
-                b.iter_batched(
-                    || {
-                        let private = ops::keypair(&mut thread_rng()).0;
-                        let mut signatures = Vec::with_capacity(n);
-                        for msg in msgs.iter() {
-                            let signature = ops::sign(&private, Some(namespace), msg);
-                            signatures.push(signature);
-                        }
-                        signatures
-                    },
-                    |signatures| {
-                        black_box(ops::aggregate(&signatures));
-                    },
-                    BatchSize::SmallInput,
-                );
-            },
-        );
+        c.bench_function(&format!("{}/msgs={}", module_path!(), msgs.len()), |b| {
+            b.iter_batched(
+                || {
+                    let private = ops::keypair(&mut thread_rng()).0;
+                    let mut signatures = Vec::with_capacity(n);
+                    for msg in msgs.iter() {
+                        let signature = ops::sign(&private, Some(namespace), msg);
+                        signatures.push(signature);
+                    }
+                    signatures
+                },
+                |signatures| {
+                    black_box(ops::aggregate(&signatures));
+                },
+                BatchSize::SmallInput,
+            );
+        });
     }
 }
