+++ storage/Cargo.toml
@@ -33,16 +33,6 @@ criterion = { workspace = true }
 bench = false
 
 [[bench]]
-name="hashmap_iteration"
+name="archive"
 harness = false
-path = "src/archive/benches/hashmap_iteration.rs"
-
-[[bench]]
-name="hashmap_load"
-harness = false
-path = "src/archive/benches/hashmap_load.rs"
-
-[[bench]]
-name="hashmap_load_fixed"
-harness = false
-path = "src/archive/benches/hashmap_load_fixed.rs"
+path = "src/archive/benches/bench.rs"
