+++ storage/src/archive/benches/hashmap_iteration.rs
@@ -12,7 +12,7 @@ struct MockIndex {
 fn benchmark_hashmap_iteration(c: &mut Criterion) {
     for n in [100_000, 1_000_000, 10_000_000] {
         for k in [4, 8, 16, 32] {
-            c.bench_function(&format!("iteration: n={} k={}", n, k), |b| {
+            c.bench_function(&format!("{}/n={} k={}", module_path!(), n, k), |b| {
                 b.iter_batched(
                     || {
                         let mut map = HashMap::with_capacity(n);
