+++ storage/src/archive/benches/hashmap_load_fixed.rs
@@ -11,7 +11,7 @@ struct MockIndex {
 
 fn benchmark_hashmap_load_fixed(c: &mut Criterion) {
     for n in [100_000, 1_000_000, 10_000_000] {
-        c.bench_function(&format!("load_fixed: n={} k={}", n, 4), |b| {
+        c.bench_function(&format!("{}/n={} k={}", module_path!(), n, 4), |b| {
             b.iter_batched(
                 || {
                     // Perform all random ops
