+++ stream/src/public_key/benches/receiver_receive.rs
@@ -4,39 +4,12 @@ use commonware_stream::{public_key::Connection, Receiver, Sender};
 use criterion::{criterion_group, criterion_main, BatchSize, Criterion};
 use futures::executor::block_on;
 
-fn benchmark_sender(c: &mut Criterion) {
-    let kbs = [1, 4, 16, 256, 1024, 4_096, 16_384, 65_536];
-    for &kb in &kbs {
-        let message_size = kb * 1024;
-        let msg = vec![0u8; message_size];
-        c.bench_function(&format!("sender: len={}", message_size), |b| {
-            b.iter_batched(
-                || {
-                    let cipher = ChaCha20Poly1305::new(&[0u8; 32].into());
-                    let (sink, stream) = mocks::Channel::init();
-                    let connection =
-                        Connection::from_preestablished(true, sink, stream, cipher, message_size);
-                    let (sender, _receiver) = connection.split();
-                    let msg = msg.clone();
-                    (sender, msg)
-                },
-                |(mut sender, msg)| {
-                    block_on(async move {
-                        sender.send(&msg).await.unwrap();
-                    });
-                },
-                BatchSize::SmallInput,
-            );
-        });
-    }
-}
-
 fn benchmark_receiver(c: &mut Criterion) {
     let kbs = [1, 4, 16, 256, 1024, 4_096, 16_384, 65_536];
     for &kb in &kbs {
         let message_size = kb * 1024;
         let msg = vec![0u8; message_size];
-        c.bench_function(&format!("receiver: len={}", message_size), |b| {
+        c.bench_function(&format!("{}/len={}", module_path!(), message_size), |b| {
             b.iter_batched(
                 || {
                     // Set up a connection between two parties.
@@ -78,5 +51,5 @@ fn benchmark_receiver(c: &mut Criterion) {
     }
 }
 
-criterion_group!(benches, benchmark_sender, benchmark_receiver);
+criterion_group!(benches, benchmark_receiver);
 criterion_main!(benches);
